+++ recipes/source.rb
@@ -102,6 +102,31 @@ when 'bluepill'
   bluepill_service 'prometheus' do
     action [:enable, :load]
   end
+when 'systemd'
+  # rubocop:disable Style/HashSyntax
+  dist_dir, conf_dir, env_file = value_for_platform_family(
+    ['fedora'] => %w(fedora sysconfig prometheus),
+    ['rhel'] => %w(redhat sysconfig prometheus)
+  )
+
+  template '/etc/systemd/system/prometheus.service' do
+    source 'systemd/prometheus.service.erb'
+    mode 0644
+    variables(:sysconfig_file => "/etc/#{conf_dir}/#{env_file}")
+    notifies :restart, 'service[prometheus]', :delayed
+  end
+
+  template "/etc/#{conf_dir}/#{env_file}" do
+    source "#{dist_dir}/#{conf_dir}/prometheus.erb"
+    mode 0644
+    notifies :restart, 'service[prometheus]', :delayed
+  end
+
+  service 'prometheus' do
+    supports :status => true, :restart => true
+    action [:enable, :start]
+  end
+  # rubocop:enable Style/HashSyntax
 else
   template '/etc/init.d/prometheus' do
     source 'prometheus.erb'